backend/requirement.txt
asgiref

djangorestframework-simplejwt
PyJWT
pytz

psycopg2-binary
python-dotenv